backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pillow
pypdf
piexif
mutagen
python-docx
openpyxl